preparing for pip package -- make lens subdirectory for python code
