+++ lens/agent/stub.py
