+++ lens/environment/condition/look_up_tables/transport_concentrations/minimal_minus_oxygen.tsv
