+++ lens/environment/condition/look_up_tables/transport_concentrations/minimal_plus_amino_acids.tsv
