+++ lens/environment/condition/look_up_tables/transport_fluxes/minimal_minus_oxygen.tsv
