+++ lens/environment/condition/look_up_tables/transport_fluxes/minimal_plus_amino_acids.tsv
