+++ lens/environment/condition/media/M9_GLC.tsv
