+++ lens/environment/condition/media_recipes.tsv
