+++ lens/environment/surrogates/metabolism.py
@@ -3,12 +3,11 @@ from __future__ import absolute_import, division, print_function
 import time
 import os
 import csv
-import numpy as np
 from scipy import constants
 
-from utils import units
-from utils.modular_fba import FluxBalanceAnalysis
-from agent.inner import CellSimulation
+from lens.utils import units
+from lens.utils.modular_fba import FluxBalanceAnalysis
+from lens.agent import CellSimulation
 
 TSV_DIALECT = csv.excel_tab
 EXTERNAL_MOLECULES_FILE = os.path.join('environment', 'condition', 'environment_molecules.tsv')
@@ -112,7 +111,7 @@ class Metabolism(CellSimulation):
 
         # convert transport limits from mmol/gDCW/hr to mol/L
         coefficient = self.dry_mass / self.cell_mass * self.cell_density * (self.timestep * units.s)  # coeff is in g*s/L
-        transport_limits_molar = {mol_id: (limit * coefficient).asNumber(units.mol/units.L)
+        transport_limits_molar = {mol_id: (limit * coefficient).asNumber(units.mol / units.L)
                                   for mol_id, limit in transport_limits.iteritems()}
 
         fba = FluxBalanceAnalysis(
