+++ lens/reconstruction/kinetic_rate_laws/parameters/example_parameters_3.json
