+++ lens/reconstruction/spreadsheets.py
@@ -8,7 +8,6 @@ import json
 import re
 import numpy as np
 
-from utils import units
 
 def array_to_list(value):
     if isinstance(value, np.ndarray):
